// src/app/interfaces/files.interface.ts
// Generated by https://quicktype.io

export interface IFileInterpreterListResponse {
  data:  IFileInterpreter[];
  total: number;
}

export interface IFileInterpreter {
  status:      boolean;
  userCreate:  string;
  createAt:    string;
  id:          string;
  url:         string;
  name:        string;
  nameFirebase: string;
  type:        string;
  description: string;
}